+++ Module.php
@@ -8,11 +8,10 @@
 
 namespace humhub\modules\transition;
 
-use humhub\libs\DynamicConfig;
+use humhub\helpers\ThemeHelper;
 use humhub\modules\content\components\ContentContainerModule;
 use humhub\modules\content\components\ContentContainerModuleManager;
 use humhub\modules\transition\jobs\SyncAllSpaceHosts;
-use humhub\modules\ui\view\helpers\ThemeHelper;
 use humhub\modules\user\models\User;
 use Yii;
 
@@ -24,11 +23,6 @@ class Module extends ContentContainerModule
      * @var string defines the icon
      */
     public $icon = 'eye';
-
-    /**
-     * @var string defines path for resources, including the screenshots' path for the marketplace
-     */
-    public $resourcesPath = 'resources';
 
     /**
      * @var int Group ID for the "Space hosts" group (for spaces' administrators and moderators)
@@ -63,7 +57,7 @@ class Module extends ContentContainerModule
      */
     public function enable()
     {
-        if (parent::enable()) {
+        if (parent::enable() !== false) {
             $this->enableTheme();
             ContentContainerModuleManager::setDefaultState(User::class, 'transition', 1);
             Yii::$app->queue->push(new SyncAllSpaceHosts());
@@ -87,7 +81,6 @@ class Module extends ContentContainerModule
         $theme = ThemeHelper::getThemeByName(self::THEME_NAME);
         if ($theme !== null) {
             $theme->activate();
-            DynamicConfig::rewrite();
         }
     }