+++ messages/de/config.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'Choose a space for each region. When a user registers a new account, depending on the region he choose, he will become a member to the corresponding space.' => 'Wählen Sie für jede Region ein Feld aus. Wenn ein Benutzer ein neues Konto registriert, wird er abhängig von der von ihm gewählten Region Mitglied des entsprechenden Bereichs.',
+  'Choose a space for each region. When a user registers a new account, depending on the region he choose, he will become a member to the corresponding space.' => 'Wählen Sie für jede Region ein Feld aus. Wenn ein Benutzer ein neues Konto registriert, wird er abhängig von der von ihm gewählten Region Mitglied des entsprechenden Space.',
   'Default space' => 'Standardraum',
   'Default spaces' => 'Standard-Leerzeichen',
   'Region' => 'Region',