+++ themes/Transition/views/calendar/widgets/wallEntry.php
@@ -13,7 +13,7 @@ use humhub\modules\calendar\widgets\EntryParticipants;
 use humhub\modules\content\widgets\richtext\RichText;
 use humhub\modules\file\widgets\FilePreview;
 use humhub\modules\ui\icon\widgets\Icon;
-use humhub\widgets\Label;
+use humhub\widgets\bootstrap\Badge;
 use yii\web\View;
 
 /* @var $calendarEntry CalendarEntry */
@@ -24,7 +24,7 @@ use yii\web\View;
 $color = $this->theme->variable('text-color-secondary');
 ?>
 
-<div class="media event calendar-wall-entry" style="margin-top:20px;" data-action-component="calendar.CalendarEntry"
+<div class="d-flex event calendar-wall-entry" style="margin-top:20px;" data-action-component="calendar.CalendarEntry"
      data-calendar-entry="<?= $calendarEntry->id ?>">
     <div class="event-info-section clearfix" style="margin-bottom:10px">
         <?= Icon::get('file-text')->color($color)->left()->size(Icon::SIZE_LG)->style('margin-top:2px;')->fixedWith() ?>
@@ -48,7 +48,7 @@ $color = $this->theme->variable('text-color-secondary');
 
                     $startTime = $formatter->getFormattedStartTime();
                     $endTime = $formatter->getFormattedEndTime();
-                    $endTime .= ' ' . Label::defaultType(Yii::$app->formatter->timeZone)->sm();
+                    $endTime .= ' ' . Badge::light(Yii::$app->formatter->timeZone)->sm();
 
                     $start = $calendarEntry->getStartDateTime();
                     $end = $calendarEntry->getEndDateTime();
@@ -76,7 +76,7 @@ $color = $this->theme->variable('text-color-secondary');
                     <small>(<?= Yii::t('CalendarModule.base', 'Recurring') ?>)</small>
                 <?php endif; ?>
                 <?php if ($calendarEntry->closed) : ?>
-                    &nbsp;<?= Label::danger(Yii::t('CalendarModule.base', 'canceled')) ?>
+                    &nbsp;<?= Badge::danger(Yii::t('CalendarModule.base', 'canceled')) ?>
                 <?php endif; ?>
             </h1>
         </div>