+++ views/admin/default-spaces.php
@@ -6,10 +6,10 @@
  * @author [Marc FARRE](https://marc.fun) for [Transition Movement](https://transitionnetwork.org/)
  */
 
-use humhub\libs\Html;
+use humhub\components\View;
+use humhub\helpers\Html;
 use humhub\modules\admin\widgets\AdminMenu;
 use humhub\modules\space\widgets\SpacePickerField;
-use humhub\modules\ui\view\components\View;
 use yii\helpers\BaseInflector;
 
 /**
@@ -26,7 +26,7 @@ AdminMenu::markAsActive(['/admin/default-spaces']);
     <h4>
         <?= $title ?>
     </h4>
-    <div class="help-block"><?= Yii::t('TransitionModule.config', 'Choose a space for each region. When a user registers a new account, depending on the region he choose, he will become a member to the corresponding space.') ?></div>
+    <div class="text-body-secondary"><?= Yii::t('TransitionModule.config', 'Choose a space for each region. When a user registers a new account, depending on the region he choose, he will become a member to the corresponding space.') ?></div>
 
     <?= Html::beginForm() ?>
     <table class="table table-striped table-hover">