+++ views/embed/calendar.php
@@ -11,10 +11,10 @@
  */
 
 use humhub\assets\AppAsset;
+use humhub\components\View;
 use humhub\modules\calendar\helpers\Url;
 use humhub\modules\calendar\widgets\FullCalendar;
 use humhub\modules\cleanTheme\assets\CleanThemeAsset;
-use humhub\modules\ui\view\components\View;
 
 AppAsset::register($this);
 CleanThemeAsset::register($this);